Update backlinks.tid
I first thought of the syntax is similar to creating a link. I didn't thought of it as a filter becasue it is written to be "input	a selection of titles".  It is better to state that it is a filter so the newbie like me won't get confused. Or editing the example as I suggested.